+++ android-project/MakaseteChoice/app/src/main/res/values/strings.xml
@@ -5,4 +5,6 @@
     <string name="save">Save</string>
     <string name="minus">-</string>
     <string name="add">+</string>
+    <string name="group_text">組み分け数</string>
+    <string name="grouping_button_text">組み分け</string>
 </resources>
